создал конструктор, рефакторинг
